Experimental POST parsing for python 3.0 and 3.1
This is an ugly workaround for broken cgi.FieldStorage implementations. DO NOT USE Python3+Bottle in production environments!
With python2.x this patch should changes nothing.
With python3.x bottle.response.POST now returns a dict with string (unicode) keys and string values. To recieve the original
bytes sent by the client, encode these with ISO-8859-1. So, if your expect UTF8 encoded data, first encode with
ISO-8859-1 and then decode with UTF8 again.

I don't know if this works in the wild. My test cases are kind of theoretical. Be warned!