Another 5% code coverage, some real and many cosmetic changes. Read more...
Added app_push() and app_pop() to easily work on a separate bottle app without interfering with other modules. This makes testing easier a lot.

Response.charset now gets its value from the Content-Type header and is read only. Add '; charset=...' to response.content_type to change it.

The static_file() helper does not change the response.header dict anymore. This side effect is gone. The new headers are part of the returned HTTPResponse object and applied later. This way it is easier to recover from an 40x error by just ignoring the returned HTTPError object. You don't have to undo the header changes anymore.

Secure cookies now work with Python3 too. This is a pure hack. Don't use it.