Bug: when using @view, returning an error response results in exception
The @view wrapper expects a dictionary of vars. It
tries to update the defaults with the returned values,
and this fails for errors.

The bottle.py patch checks and only updates and calls
the template if the result is a dict.

The app.py patch raises errors instead of using return, and
so escapes the view processing.

Missing: test cases