Fix templates so they can handle code with non-ascii characters inside
The simple templates break when a template file has code with non-ascii
characters in it, like:

        {{"ñaca"}}

The cause is that PyStmt.__repr__ returns an unicode object that is not
ascii-encodable, as required by Python's repr() implementation.

This patch fixes it by making PyStmt.__repr__ return an utf-8 encoded string,
and making flush() decoding it back before using it.

It also adds a testcase for this scenario to help prevent similar issues in
the future.