fix: (issue #85) fapws3 segfaults if port is an integer
WARNING: Old releases of fapws3 are no longer supported and may segfault!

There is no way to detect the version of fapws3 or the expected type of the port argument.
The wrong type causes segfaults. If you are forced to use an old release of fapws3, configure the
server manually.