fix: (issue #90) Reloading server dies on slow hardware.
Python seems to have some serious issues with the GIL when IO bound and CPU
bound threads are mixed on a multi-core CPU [1]. This may prevent the observer
thread to wake up and touch the lock-file it time. I observed this even on
otherwise idling hardware. This patch does not fix the bug completely, but it
should give the observer enough time and work around this issue in most cases.
Dying observer processes are relatively rare. An increased timeout should not
hurt that much.

[1] http://dabeaz.blogspot.com/2010/01/python-gil-visualized.html