Security fix: cookie_decode() vulnerability to statistical timing attacks.
This vulnerability reduces the amount of time required to guess a valid
signature for a forged cookie. Only applications that use the secure-cookie
feature are affected.

Detail: Prior to this patch, the transmitted signature and the calculated
(valid) signature were compared using the python '==' operator. The runtime of
this operation depends on the length of a common string prefix. The time
difference is barely measurable for a single request, but it exists. An attacker
could send a large amounts of request with different signatures, measure the
time for each request and statistically determine which signature is most likely
to have a valid prefix. On a fast network, this can drastically reduce the time
required to guess a valid signature for a forged cookie.