Reduced module load time by ~30% (~20ms per CGI or GAE request) and improved template parser performance.
As I found out today, 'tokenize' is a VERY hervy module. Importing it takes
about 22ms on a modern CPU. This module alone was responsible for 25% of
bottles load time. To get rid of it, I had to remove the 'inspect' dependency
as well (which uses 'tokenize' internally). The only remaining piece of code
that depends on 'inspect' is 'yieldroutes()', so try to avoid it in GAE or CGI
environments.

Bottle used 'tokenize' to strip comments from stpl templates. I solved this
with a 4-line regular expression. As a side effect, template parsing is up to
30% faster now. It's even less code than before. Regular expressions are fun :)