Added an 'id' field to the route context dictionary. Added Bottle.reset() to allow plugins to refresh a route at runtime.
A plugin can now call context['app'].reset(context['id']) to get a chance
to re-apply or remove itself to or from a route. Example: If the first 1000
requests don't return a dict, the JSONPlugin could remove itself from that
route by adding itself to the context['skip'] list.

This kind of dynamic behaviour is very hard to debug. Use it with care.