Added Request.urlparts property and lots of redirect() tests.
This reverts 5cf06439. The result of Request.url is now quoted properly.
If you really need the URL with an unquoted path, get Request.urlparts,
urllib.unquote() the path and urlparse.urlunsplit() it again.

I do not quite remember why I changed the Request.url behavior in the
first place. An URL with special characters in the path is useless.