Fix: PATH_INFO always contains "/" for mounted app root
When application is mounted to `/some_path` then `PATH_INFO` contains `/` for
requests to both `/some_path` and `/some_path/`. So it is probably impossible
to distinguish between these urls in mounted application (e.g. to redirect to
normalized url).

According to wsgi spec `PATH_INFO` __may__ be empty in this case. It's seems
to be possible to get desired behaviour by changing `path_shift()`.