Fix #387: Template lookup found files in workdir, even if TEMPLATE_PATH did not contain '.'.
The template path list is now obeyed (with an exception for templates with an explicit absolute path).

The next release will tighten the rules a bit further: TEMPLATE_PATH must not be empty, and absolute paths are
considered relative, too. This ensures that only templates from within TEMPLATE_PATH directries are loaded.