static_file: add charset to content-type
currently static_file serves content-type as 'text/html' without
providing a charset, which makes firefox bark with this in the error log:

"""
The character encoding of the HTML document was not declared. The document
will render with garbled text in some browser configurations if the
document contains characters from outside the US-ASCII range. The
character encoding of the page must be declared in the document or in the
transfer protocol.
"""

this patch modifies static_file to add "charset=UTF-8" to all "text/..."
mimetypes in case the mimetype was not provided explicitly. The actual
charset used can be configured via 'autocharset' keyword argument.