Fixed doc translation workflow.
The update.sh script is useful as long as sphinx-intl is not part of debian.
The makefile had an shell string bug. '$var' != "$var"
Sphinx doctrees should not be part of the git repository.
And there was a print left in docs/conf.py as well as an import error.
Importing the correct version of bottle (and not the system installed one) is important
 to get the correct docstrings for the API documentation.