Improved ConfigDict.load_config() documentation and functionality.
The issue #784 criticized that values under the [DEFAULT] section are injected into all other sections while the documentation states that these values are only injected into the root namespace.

The actual behavior is correct and the documentation was wrong. See configparser.ConfigParser() documentation for the exact behavior. There is no easy way to change this behavior.

This patch:
  - fixes #784: ConfigDict.load_config() and 'DEFAULT' sections.
  - Improves documentation.
  - Enables extended value string interpolation for Python 3.
  - Enables empty (none) values.
  - Enables customisation by passing configparser.ConfigParser() parameters to ConfigDict.load_config()
  - Introduces the special [ROOT] section as an replacement for [DEFAULT].