Explicitly add global variables to the globals dictionary passed to exec statement when executing compiled templates.
SimpleTemeplate is implemented as a compiled python code object executed using the builtin function exec(). The global variables visible in the compiled statement are passed as a dictionary. This dictionary now contains explicitly all definitions of the global context, in which exec() is called. Otherwise, specifying a variable in the URL pattern called __builtins__ leads to an empty globals() context, because exec() would not add the globals() definitions implicitly in that case. All calls of builtins in the compiled statement would lead to an exception.
See https://docs.python.org/3/library/functions.html#exec for the implicit definition of necessary global context.