Fix for proper decoding non-latin1 symbols
If there are cyrillic or other non-latin1 symbols in data in request and there are simultaneously files in request there will be UnicodeEncodeError because these symbols cannot be encoded in latin1. These error appears only if there are files in request, e.g. content-type header is multipart/form-data. If there are no files in request and content-type header is application/x-www-form-urlencoded then will be no UnicodeEncodeError (cyrillic symbols in data decoded properly).