Offer option to add custom headers in static_file
static_file does not offer the ability of adding custom headers because the headers dictionary is built inside the function.
changed it so it is a parameter with a default value. Behavior is not changed and the user has the option of supplying custom headers to the response if needed.