Add support for non-empty requests with no Content-Length header.
In some situations, a non-empty non-chunked request may lack a
Content-Length header and still be valid:
- HTTP/1.0 clients (default to 'Connection: close')
- HTTP/1.1 clients with 'Connection: close'
- WSGI server transparently decoding 'Transfer-Encoding: chunked'
- WSGI middleware filtering the request body (e.g. gzip)

In all these situations, 'wsgi.input' MUST NOT read across
request boundaries and terminate at the end of the request body,
because WSGI defines no other way to detect the end of the stream.
Most WSGI servers or filtering middleware correctly set 'wsgi.input'
to a limiting wrapper. They MAY signal this via the non-standard
'wsgi.input_terminated' flag, but since not limiting 'wsgi.input'
would be clearly wrong, we can assume support for 'wsgi.input_terminated'
(or 'Connection: close', which has the same effect) even if the flag
is not set.

This patch will cause bottle to read from 'wsgi.input' until it returns
no more bytes if 'Content-Length' is not set. It will also throw an error
if 'Content-Length' is set, but less than the expected number of bytes can
be read.