Add tests for #602 bug
It looks that #602 is gone, but I'm not sure when exactly it was fixed.

I see that there is a suspicious change [here](https://github.com/bottlepy/bottle/commit/d85a6983ceedacd9ab9afbcd027139d8773b67ac):

    environ['PATH_INFO'] = path.encode('latin1').decode('utf8', 'ignore')

Where original cause of the error was removed and replaced by `'ignore'`.

Also fixed deprecation warnings, where:

    @self.subapp.route('')

Does exactly same thing as this:

    @self.subapp.route('/test/<test>')

Where `''` is passed to `route()`, then `makelist()` utility assumes that nothing is passed and falls back to path autogeneration from callbackf unction. Not sure if this is expected behaviour?