fixes #580 and makes it "fit"
This fixes #580 

It seems the current release and the dev versions are different.  This dev version adds a headers kwarg, which is not how the rest of bottle works, which uses `response.headers()`.  This fixes #580, which was already, somewhat fixed, but also changes it to not need a headers kwarg, just copy from the existing response headers.

Either way, you can close #580 now.