fix: Replace cgi.FieldStorage
Standard library cgi.FieldStorage was deprecated in Python 3.11 and removed in 3.13.
We now have to ship our own multipart form data parser implementation, which is a port
of the existing 'multipart' module. The new parser is used for all python versions,
which may break backwards compatibility for certain edge cases and error scenarios.
For a 'good' request there should be no difference, though.

What changed:
* Lax newline parsing (accepting \r or \n in place of \r\n) is a security risk and no longer implemented.
  Most modern browsers and client libraries follow the spec and should not have any issues producing valid multipart data.
* Parsing errors are no longer silently ignored, but trigger an appropiate 400 error.