change: Rename `bottle.py` script install to `bottle` docs: Update details on CLI use, release notes, and minor other changes
Naming the script command `bottle.py` breaks with a circular import
error because that script tries to `import bottle` and finds itself on
the module search path instead of the actual bottle module.

This worked with setuptools because instead of as small entrypoint
script, setuptools generated wheels copied the entire contents of
bottle.py into the bin/Scripts path of the environment (and bottle fixed
sys.modules at runtime). This is no longer feasible with modern
pyproject.toml script entrypoints and the only sensible way moving
forward is to rename the `bottle.py` script entrypoint to just `bottle`
and deprecate the old name in Bottle 0.13.

Hint: You can still execute bottle.py directly or via `python -m bottle`.
The only change is that the command installed by pip into the bin/Scripts
folder of the (virtual) environment is now called `bottle`.