fix: Workaround for uv/uvx bug
The uv/uvx tools break scripts that contain a __future__ import by
injecting a multiline string infront of it. Future imports are
very limited on where they can appear.
See https://github.com/astral-sh/uv/issues/6489

We cannot remove the __future__ import in bottle-0.13 because Python 2.7 support
still depends on it, but we can move it above the docstring, so the uv/uvx
modifications no longer trigger a SyntaxError.

fix #1486