Remove automatic monkey-patching for gevent based servers.
Bottle used to depend on monkey-patching to work with greenlet-based concurrency.  This is no longer the case, so we can remove this error-prone and incomplete workaround from the bottle CLI startup logic.

change: Bottle CLI no longer monkey-patches the python standard library when selecting the gevent or eventlet server adapters from the command line.